requirements.txt
streamlit
openai>=1.23
requests
pillow
python-dotenv
